chore(ordered-list): remove extra stories
