+++ packages/react/src/components/OrderedList/OrderedList.mdx
@@ -1,4 +1,4 @@
-import { Props } from '@storybook/addon-docs';
+import { Props, Preview, Story } from '@storybook/addon-docs';
 
 # OrderedList
 
@@ -11,6 +11,10 @@ import { Props } from '@storybook/addon-docs';
 ## Table of Contents
 
 ## Overview
+
+<Preview>
+  <Story id="components-orderedlist--playground" />
+</Preview>
 
 ## Component API
