+++ packages/react/src/components/OrderedList/OrderedList.stories.js
@@ -1,0 +1,84 @@
+import React from 'react';
+import OrderedList from '../OrderedList';
+import ListItem from '../ListItem';
+import mdx from './OrderedList.mdx';
+
+export default {
+  title: 'Components/OrderedList',
+  component: OrderedList,
+  subcomponents: {
+    ListItem,
+  },
+  parameters: {
+    docs: {
+      page: mdx,
+    },
+  },
+};
+
+export const Default = () => (
+  <OrderedList>
+    <ListItem>Ordered List level 1</ListItem>
+    <ListItem>Ordered List level 1</ListItem>
+    <ListItem>Ordered List level 1</ListItem>
+    <ListItem>Ordered List level 1</ListItem>
+    <ListItem>Ordered List level 1</ListItem>
+    <ListItem>Ordered List level 1</ListItem>
+    <ListItem>Ordered List level 1</ListItem>
+    <ListItem>Ordered List level 1</ListItem>
+    <ListItem>Ordered List level 1</ListItem>
+    <ListItem>Ordered List level 1</ListItem>
+    <ListItem>Ordered List level 1</ListItem>
+    <ListItem>Ordered List level 1</ListItem>
+    <ListItem>Ordered List level 1</ListItem>
+  </OrderedList>
+);
+
+export const Playground = (args) => (
+  <OrderedList {...args}>
+    <ListItem>Ordered List level 1</ListItem>
+    <ListItem>Ordered List level 1</ListItem>
+    <ListItem>Ordered List level 1</ListItem>
+    <ListItem>Ordered List level 1</ListItem>
+    <ListItem>Ordered List level 1</ListItem>
+    <ListItem>Ordered List level 1</ListItem>
+    <ListItem>Ordered List level 1</ListItem>
+    <ListItem>Ordered List level 1</ListItem>
+    <ListItem>Ordered List level 1</ListItem>
+    <ListItem>Ordered List level 1</ListItem>
+    <ListItem>Ordered List level 1</ListItem>
+    <ListItem>Ordered List level 1</ListItem>
+    <ListItem>Ordered List level 1</ListItem>
+  </OrderedList>
+);
+
+Playground.argTypes = {
+  children: {
+    table: {
+      disable: true,
+    },
+  },
+  className: {
+    table: {
+      disable: true,
+    },
+  },
+  isExpressive: {
+    control: {
+      type: 'boolean',
+    },
+    defaultValue: false,
+  },
+  native: {
+    control: {
+      type: 'boolean',
+    },
+    defaultValue: false,
+  },
+  nested: {
+    control: {
+      type: 'boolean',
+    },
+    defaultValue: false,
+  },
+};
